+++ src/components/app/app.jsx
@@ -1,6 +1,6 @@
 import React, {PureComponent} from "react";
 import {BrowserRouter} from "react-router-dom";
-import Screens from "../screens";
+import Screens from "../screens/importComponent";
 
 class App extends PureComponent {
   render() {
